make StaticArrays dependency into an extension
