+++ src/ForwardDiff.jl
@@ -1,8 +1,7 @@
 module ForwardDiff
 
 using DiffRules, DiffResults
-using DiffResults: DiffResult, MutableDiffResult, ImmutableDiffResult
-using StaticArrays
+using DiffResults: DiffResult, MutableDiffResult
 using Preferences
 using Random
 using LinearAlgebra
@@ -22,6 +21,10 @@ include("derivative.jl")
 include("gradient.jl")
 include("jacobian.jl")
 include("hessian.jl")
+
+if !isdefined(Base, :get_extension)
+    include("../ext/StaticArraysExt.jl")
+end
 
 export DiffResults
